test/web-platform-tests/to-upstream/css/cssom/style-backgroundimage-lineargradient.html

<!DOCTYPE html>
<title>CSSOM: Setting style.backgroundColor to oklch()</title>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://drafts.csswg.org/cssom/#the-elementcssinlinestyle-mixin">
<!-- Regression test for https://github.com/jsdom/jsdom/issues/3552 -->

<div id="test"></div>

<script>
"use strict";

test(() => {
  const div = document.getElementById("test");

  div.style.backgroundImage = "linear-gradient(green, blue)";
  assert_equals(div.style.backgroundImage, "linear-gradient(green, blue)");

}, "Setting style.backgroundImage using linear-gradient()");

</script>
